[refactor]: cookie message 확인 useEffect 위치 변경
